requirements.txt
asgiref==3.8.1
attrs==24.3.0
cffi==1.17.1
cryptography==44.0.0
Django==5.1.5
django-cors-headers==4.6.0
django-filter==24.3
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Markdown==3.7
phonenumbers==8.13.52
pillow==11.1.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.22.3
sqlparse==0.5.3
tzdata==2024.2
uritemplate==4.1.1
